remove LocalScene in favor of Scene
